correções de bugs e custom hook para GET
